+++ tests/instantiation.js
@@ -86,64 +86,6 @@ export default function instantiation(Graph, implementation, checkers) {
       },
 
       /**
-       * defaultEdgeAttributes
-       */
-      'defaultEdgeAttributes': {
-
-        'providing something other than a plain object should throw.': function() {
-          assert.throws(function() {
-            const graph = new Graph({defaultEdgeAttributes: 'test'});
-          }, invalid());
-        },
-
-        'it should set default attributes on edges.': function() {
-          const graph = new Graph({defaultEdgeAttributes: {type: 'KNOWS'}, multi: true});
-
-          graph.addNodesFrom(['John', 'Martha']);
-
-          let edge = graph.addEdge('John', 'Martha');
-
-          assert.deepEqual(graph.getEdgeAttributes(edge), {type: 'KNOWS'});
-
-          edge = graph.addEdge('John', 'Martha', {weight: 3});
-
-          assert.deepEqual(graph.getEdgeAttributes(edge), {weight: 3, type: 'KNOWS'});
-
-          edge = graph.addEdge('John', 'Martha', {type: 'LIKES'});
-
-          assert.deepEqual(graph.getEdgeAttributes(edge), {type: 'LIKES'});
-        }
-      },
-
-      /**
-       * defaultNodeAttributes
-       */
-      'defaultNodeAttributes': {
-
-        'providing something other than a plain object should throw.': function() {
-          assert.throws(function() {
-            const graph = new Graph({defaultNodeAttributes: 'test'});
-          }, invalid());
-        },
-
-        'it should set default attributes on nodes.': function() {
-          const graph = new Graph({defaultNodeAttributes: {eyes: 'blue'}});
-
-          graph.addNode('John');
-
-          assert.deepEqual(graph.getNodeAttributes('John'), {eyes: 'blue'});
-
-          graph.addNode('Thomas', {age: 23});
-
-          assert.deepEqual(graph.getNodeAttributes('Thomas'), {age: 23, eyes: 'blue'});
-
-          graph.addNode('Martha', {eyes: 'green'});
-
-          assert.deepEqual(graph.getNodeAttributes('Martha'), {eyes: 'green'});
-        }
-      },
-
-      /**
        * edgeKeyGenerator
        */
       'edgeKeyGenerator': {